IMP - Refactor path definitions
